feat: extract debug and snackbar helpers
