+++ src/helpers/classicBattle/uiEventHandlers.js
@@ -3,15 +3,11 @@ import { getOpponentCardData } from "./opponentController.js";
 import * as scoreboard from "../setupScoreboard.js";
 import { showSnackbar } from "../showSnackbar.js";
 import { t } from "../i18n.js";
-import {
-  renderOpponentCard,
-  showRoundOutcome,
-  showStatComparison,
-  updateDebugPanel
-} from "./uiHelpers.js";
+import { renderOpponentCard, showRoundOutcome, showStatComparison } from "./uiHelpers.js";
+import { updateDebugPanel } from "./debugPanel.js";
+import { getOpponentDelay } from "./snackbar.js";
 
 let opponentSnackbarId = 0;
-let opponentDelayMs = 500;
 
 export function bindUIHelperEventHandlersDynamic() {
   // Ensure we only bind once per EventTarget instance
@@ -39,7 +35,7 @@ export function bindUIHelperEventHandlersDynamic() {
       if (!opts.delayOpponentMessage) {
         opponentSnackbarId = setTimeout(
           () => showSnackbar(t("ui.opponentChoosing")),
-          opponentDelayMs
+          getOpponentDelay()
         );
       }
     } catch {}
