+++ tests/helpers/classicBattle/cooldown.skipHandlerReady.test.js
@@ -28,7 +28,7 @@ describe("skip handler clears fallback timer", () => {
       showSnackbar: vi.fn(),
       updateSnackbar: vi.fn()
     }));
-    vi.doMock("../../../src/helpers/classicBattle/uiHelpers.js", () => ({
+    vi.doMock("../../../src/helpers/classicBattle/debugPanel.js", () => ({
       updateDebugPanel: vi.fn()
     }));
     vi.doMock("../../../src/helpers/classicBattle/orchestrator.js", () => ({
