+++ tests/helpers/classicBattle/onTransition.helpers.test.js
@@ -13,7 +13,7 @@ vi.mock("../../../src/helpers/setupScoreboard.js", () => ({
   clearMessage: vi.fn(),
   showMessage: vi.fn()
 }));
-vi.mock("../../../src/helpers/classicBattle/uiHelpers.js", () => ({
+vi.mock("../../../src/helpers/classicBattle/debugPanel.js", () => ({
   updateDebugPanel: vi.fn()
 }));
