+++ tests/helpers/classicBattle/timerService.drift.test.js
@@ -49,7 +49,9 @@ describe("timerService drift handling", () => {
     }));
     vi.doMock("../../../src/helpers/classicBattle/uiHelpers.js", () => ({
       enableNextRoundButton: vi.fn(),
-      disableNextRoundButton: vi.fn(),
+      disableNextRoundButton: vi.fn()
+    }));
+    vi.doMock("../../../src/helpers/classicBattle/debugPanel.js", () => ({
       updateDebugPanel: vi.fn()
     }));
     const cool = createDriftStarter();
